+++ gameInfos/extractGameInfos.py
@@ -3,6 +3,7 @@ import os
 
 BASE_PATH = os.path.expandvars("%LOCALAPPDATA%low\\tobspr Games\\shapez 2\\")
 BUILDINGS_PATH = BASE_PATH + "buildings-metadata.json"
+ADDITIONAL_BUILDINGS_PATH = "./gameInfos/additionalBuildings.json"
 RESEARCH_PATH = BASE_PATH + "research-metadata-full.json"
 EXTRACTED_BUILDINGS_PATH = "./gameInfos/buildings.json"
 EXTRACTED_RESEARCH_PATH = "./gameInfos/research.json"
@@ -43,34 +44,16 @@ def main() -> None:
 
 
     # buildings
-    TO_REMOVE_BUILDINGS = ["HubBuilding"]
-    TO_ADD_BUILDINGS = [
-        {
-            "Id" : "HubBuilding",
-            "Title" : "Hub",
-            "Variants" : [
-                {
-                    "Id" : "HubDefaultVariant",
-                    "Title" : "Hub",
-                    "InternalVariants" : [
-                        {
-                            "Id" : "HubDefaultInternalVariant",
-                            "Tiles" : [
-                                {}
-                            ]
-                        }
-                    ]
-                }
-            ]
-        }
-    ]
 
     with open(BUILDINGS_PATH,encoding="utf-8") as f:
         buildingsRaw = json.load(f)
+    with open(ADDITIONAL_BUILDINGS_PATH,encoding="utf-8") as f:
+        additionalBuildings = json.load(f)
+    toRemoveBuildings = [ab["Id"] for ab in additionalBuildings]
 
-    buildingsRaw = [b for b in buildingsRaw if b["Id"] not in TO_REMOVE_BUILDINGS]
+    buildingsRaw = [b for b in buildingsRaw if b["Id"] not in toRemoveBuildings]
     extractedBuildings:dict[str,str|list] = {"GameVersion":gameVersion,"Buildings":[]}
-    for variantListRaw in TO_ADD_BUILDINGS+buildingsRaw:
+    for variantListRaw in additionalBuildings+buildingsRaw:
         extractedVariantList = extractKeys(variantListRaw,{},["Id","Title"])
         extractedVariantList["Variants"] = []
         for internalVariantListRaw in variantListRaw["Variants"]: